added some and deleted some

+++ src/render/frames.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   frames.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mmitkovi <mmitkovi@student.42.fr>          +#+  +:+       +#+        */
+/*   By: hgatarek <hgatarek@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/07 13:07:28 by mmitkovi          #+#    #+#             */
-/*   Updated: 2025/10/14 14:34:55 by mmitkovi         ###   ########.fr       */
+/*   Updated: 2025/10/14 18:50:49 by hgatarek         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../includes/cub3d.h"
 
-long	get_time(void)
+double	get_time(void)
 {
 	struct timeval tv;
 	
@@ -20,14 +20,14 @@ long	get_time(void)
 	return(tv.tv_sec * 1000 + tv.tv_usec / 1000);
 }
 
-float calc_timeframe(t_data *data)
+double calc_timeframe(t_data *data)
 {
 	double now;
 	double frame_time;
 	
 	now = get_time();
 	frame_time = (now - data->last_time) / 1000.0;
-	// clamp delta time to prevent huge jumps
+	// clamp delta time to prevent huge jumps     //probably to uncomment
 	if (frame_time > 0.05f)
 		frame_time = 0.05f;
     if (frame_time < 0.001f)
